--- src/Endpoints.Test/CrudPipelineTests.cs
@@ -1,286 +1,0 @@
-using System.Threading.Tasks;
-using System.Net;
-using System.Net.Http;
-using Xunit;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Endpoints.Pipelines;
-using Microsoft.AspNetCore.Http;
-using System.IO;
-using Endpoints.Extensions;
-using Newtonsoft.Json;
-using Moq;
-using Microsoft.AspNetCore.TestHost;
-
-namespace Endpoints.Test
-{
-    public class CrudPipelineTests : IClassFixture<TestFixture>
-    {
-        private readonly TestFixture _fixture;
-
-        public CrudPipelineTests(TestFixture fixture)
-        {
-            _fixture = fixture;
-        }
-
-        private TestServer CreateTestCrudServer(IDatabase<CrudId, CrudModel> database)
-        {
-            // Arrange
-            return _fixture.CreateServer(services =>
-            {
-                services.AddSingleton(database);
-                services.AddPipelines();
-                services.RegisterPipeline<CreateCrudModel, CrudModel, CrudId>();
-                services.RegisterPipeline<ReadCrudModel, CrudId, CrudModel>();
-                services.RegisterPipeline<UpdateCrudModel, UpdateCrudModelRequest, bool>();
-                services.RegisterPipeline<DeleteCrudModel, CrudId, bool>();
-            },
-            app => app.UseEndpoints(endpoints =>
-            {
-                var registry = endpoints.ServiceProvider.GetRequiredService<PipelineRegistry>();
-                endpoints.MapPost("/model", registry.Get<CreateCrudModel, CrudModel, CrudId>());
-                endpoints.MapGet("/model/{id}", registry.Get<ReadCrudModel, CrudId, CrudModel>());
-                endpoints.MapPut("/model/{id}", registry.Get<UpdateCrudModel, UpdateCrudModelRequest, bool>());
-                endpoints.MapDelete("/model/{id}", registry.Get<DeleteCrudModel, CrudId, bool>());
-            }));
-        }
-
-        [Fact]
-        public async Task CanCreate()
-        {
-            var database = new Mock<IDatabase<CrudId, CrudModel>>();
-            database.Setup(d => d.Create(It.IsAny<CrudModel>()))
-                .Returns(Task.FromResult(new CrudId("new-model-id")));
-
-            using var server = CreateTestCrudServer(database.Object);
-            var model = new CrudModel("", "model-body");
-            var requestContent = new StringContent(JsonConvert.SerializeObject(model));
-
-            var client = server.CreateClient();
-
-            // Act
-            var response = await client.PostAsync("/model", requestContent);
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var content = await response.Content.ReadAsStringAsync();
-            Assert.Equal("new-model-id", content);
-        }
-
-        [Fact]
-        public async Task CanRead()
-        {
-            var database = new Mock<IDatabase<CrudId, CrudModel>>();
-            database.Setup(d => d.Read(It.Is<CrudId>(id => id.Id == "model-id")))
-                .Returns(Task.FromResult(new CrudModel("model-id", "model-body")));
-
-            using var server = CreateTestCrudServer(database.Object);
-
-            var client = server.CreateClient();
-
-            // Act
-            var response = await client.GetAsync("/model/model-id");
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var content = await response.Content.ReadAsStringAsync();
-            Assert.Equal(@"{""Id"":""model-id"",""Body"":""model-body""}", content);
-        }
-
-        [Fact]
-        public async Task CanUpdate()
-        {
-            var database = new Mock<IDatabase<CrudId, CrudModel>>();
-            database.Setup(d => d.Update(It.IsAny<CrudId>(), It.IsAny<CrudModel>()))
-                .Returns(Task.CompletedTask);
-
-            var model = new CrudModel("", "model-body");
-            var requestContent = new StringContent(JsonConvert.SerializeObject(model));
-
-            using var server = CreateTestCrudServer(database.Object);
-            var client = server.CreateClient();
-
-            // Act
-            var response = await client.PutAsync("/model/model-id", requestContent);
-
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var content = await response.Content.ReadAsStringAsync();
-            Assert.Equal("", content);
-        }
-
-        [Fact]
-        public async Task CanDelete()
-        {
-            var database = new Mock<IDatabase<CrudId, CrudModel>>();
-            database.Setup(d => d.Delete(It.IsAny<CrudId>()))
-                .Returns(Task.CompletedTask);
-
-            using var server = CreateTestCrudServer(database.Object);
-            var client = server.CreateClient();
-
-            // Act
-            var response = await client.DeleteAsync("/model/model-id");
-
-
-            // Assert
-            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-
-            var content = await response.Content.ReadAsStringAsync();
-            Assert.Equal("", content);
-        }
-
-        public record UpdateCrudModelRequest
-        {
-            public CrudId Id { get; set; }
-            public CrudModel Model { get; set; }
-
-            public UpdateCrudModelRequest(string id, CrudModel model) => (Id, Model) = (new CrudId(id), model);
-        }
-
-        public record CrudModel
-        {
-            public string Id { get; set; }
-            public string Body { get; set; }
-
-            public CrudModel(string id, string body) => (Id, Body) = (id, body);
-        }
-
-        public record CrudId
-        {
-            public string Id { get; set; }
-
-            public CrudId(string id) => (Id) = (id);
-        }
-
-        public interface IDatabase<TId, TModel>
-        {
-            Task<TId> Create(TModel model);
-            Task<TModel> Read(TId id);
-            Task Update(TId id, TModel model);
-            Task Delete(TId id);
-        }
-
-        public class CreateCrudModel : Pipeline<CrudModel, CrudId>
-        {
-            private readonly IDatabase<CrudId, CrudModel> _database;
-
-            public CreateCrudModel(IDatabase<CrudId, CrudModel> database)
-            {
-                _database = database;
-            }
-
-            protected override async Task<CrudId> GetResponse(CrudModel input)
-            {
-                return await _database.Create(input);
-            }
-
-            protected override CrudModel ParseModel(HttpContext context) => throw new System.NotImplementedException();
-
-            protected override async Task<CrudModel> ParseModelAsync(HttpContext context)
-            {
-                var body = await context.ParseBody();
-                return JsonConvert.DeserializeObject<CrudModel>(body);
-            }
-
-            protected override async Task ParseResponse(HttpContext context, CrudId response)
-            {
-                await context.Response.WriteAsync(response.Id);
-            }
-        }
-
-        public class ReadCrudModel : Pipeline<CrudId, CrudModel>
-        {
-            private readonly IDatabase<CrudId, CrudModel> _database;
-
-            public ReadCrudModel(IDatabase<CrudId, CrudModel> database)
-            {
-                _database = database;
-            }
-
-            protected override async Task<CrudModel> GetResponse(CrudId input)
-            {
-                return await _database.Read(input);
-            }
-
-            protected override CrudId ParseModel(HttpContext context)
-            {
-                var id = context.Request.RouteValues["id"]?.ToString();
-                return new CrudId(id);
-            }
-
-            protected override async Task ParseResponse(HttpContext context, CrudModel response)
-            {
-                await context.Response.WriteAsync(JsonConvert.SerializeObject(response));
-            }
-        }
-
-        public class UpdateCrudModel : Pipeline<UpdateCrudModelRequest, bool>
-        {
-            private readonly IDatabase<CrudId, CrudModel> _database;
-
-            public UpdateCrudModel(IDatabase<CrudId, CrudModel> database)
-            {
-                _database = database;
-            }
-
-            protected override async Task<bool> GetResponse(UpdateCrudModelRequest input)
-            {
-                await _database.Update(input.Id, input.Model);
-                return true;
-            }
-
-            protected override UpdateCrudModelRequest ParseModel(HttpContext context) => throw new System.NotImplementedException();
-
-            protected override async Task<UpdateCrudModelRequest> ParseModelAsync(HttpContext context)
-            {
-                var id = context.Request.RouteValues["id"]?.ToString();
-
-                var body = await context.ParseBody();
-                var model = JsonConvert.DeserializeObject<CrudModel>(body);
-
-                return new UpdateCrudModelRequest(id, model);
-            }
-
-            protected override Task ParseResponse(HttpContext context, bool response)
-            {
-                context.Response.StatusCode = 200;
-
-                return Task.CompletedTask;
-            }
-        }
-
-        public class DeleteCrudModel : Pipeline<CrudId, bool>
-        {
-            private readonly IDatabase<CrudId, CrudModel> _database;
-
-            public DeleteCrudModel(IDatabase<CrudId, CrudModel> database)
-            {
-                _database = database;
-            }
-
-            protected override CrudId ParseModel(HttpContext context)
-            {
-                var id = context.Request.RouteValues["id"]?.ToString();
-                return new CrudId(id);
-            }
-
-            protected override async Task<bool> GetResponse(CrudId input)
-            {
-                await _database.Delete(input);
-                return true;
-            }
-
-            protected override Task ParseResponse(HttpContext context, bool response)
-            {
-                context.Response.StatusCode = 200;
-
-                return Task.CompletedTask;
-            }
-        }
-    }
-}
