+++ src/Endpoints/Extensions/ServiceCollectionExtensions.cs
@@ -17,20 +17,6 @@ namespace Endpoints.Extensions
             return services;
         }
 
-        public static IServiceCollection RegisterPipeline<TPipeline, TIn, TOut>(
-            this IServiceCollection services,
-            Action<IPipelineBuilder<TPipeline, TIn, TOut>> builder = null)
-            where TPipeline : Pipeline<TIn, TOut>
-        {
-            var instructions = new PipelineInstructions<TPipeline, TIn, TOut>();
-            if (builder != null)
-                builder(instructions);
-
-            services.AddSingleton(instructions);
-
-            return services;
-        }
-
         public static IServiceCollection RegisterRetrievePipeline<TIn, TOut>(
            this IServiceCollection services,
            Func<HttpContext, Task<TIn>> parseModel,
@@ -105,47 +91,6 @@ namespace Endpoints.Extensions
             services.AddSingleton(instructions);
 
             return services;
-        }
-    }
-
-    public class PipelineRegistry
-    {
-        private readonly IServiceProvider _serviceProvider;
-
-        public PipelineRegistry(IServiceProvider serviceProvider)
-        {
-            _serviceProvider = serviceProvider;
-        }
-
-        public RequestDelegate Get<TPipeline, TIn, TOut>()
-            where TPipeline : Pipeline<TIn, TOut>
-        {
-            var instructions = _serviceProvider.GetRequiredService<PipelineInstructions<TPipeline, TIn, TOut>>();
-            var pipeline = instructions.GetPipeline(_serviceProvider);
-
-            return pipeline.Run;
-        }
-
-        public RequestDelegate GetRetrieve<TRetriever, TIn, TOut>()
-            where TRetriever : IRetriever<TIn, TOut>
-        {
-            var instructions = _serviceProvider.GetRequiredService<RetrievePipelineInstructions<TIn, TOut>>();
-            var (pipeline, ok) = instructions.TryGetPipeline<TRetriever, TIn, TOut>(_serviceProvider);
-            if (!ok)
-                throw new Exception("Could not create pipeline");
-
-            return pipeline.Run;
-        }
-
-        public RequestDelegate GetRetrieve<TIn, TOut>(Func<IServiceProvider, Func<TIn, Task<TOut>>> retrieverFn)
-        {
-            var instructions = _serviceProvider.GetRequiredService<RetrievePipelineInstructions<TIn, TOut>>();
-            var retriever = retrieverFn(_serviceProvider);
-            var (pipeline, ok) = instructions.TryGetPipeline<TIn, TOut>(retriever, _serviceProvider);
-            if (!ok)
-                throw new Exception("Could not create pipeline");
-
-            return pipeline.Run;
         }
     }
 }
